+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Pending release
 
-### Contracts
+### Contract changes
+
+* Update `status-message-collections` to use `LookupMap`
+* **BREAKING** Update `fungible-token` implementation to use `LookupMap`. It changes storage layout.
+
+### API changes
 
 * Introduce `LookupMap` and `LookupSet` that are faster implementations of `UnorderedMap` and `UnorderedSet`, but without support for iterators.
 Most read/lookup/write are done in 1 storage access instead of 2 or 3 for `Unordered*` implementations.